// HRMS Frontend/HRMS/src/app/models/SalaryModels/Salary-Data-model.ts
export type SalaryDataModel = {
  empName: string;
  departmentName: string;
  month : string;
  year: string;
  salary: number;
  attendance: number;
  absence: number;
  specialSettingsBonus: number;
  specialSettingsPenality: number;
  totalBouns: number;
  totalPenality: number;
  totalSalary: number;
};